scripts/ocr/requirements.txt
rapidocr-onnxruntime
opencv-python-headless
Pillow
